reorganize router code, add NetlistRouter
also: clip high-fanout BBs by original BBs to avoid
data races in the parallel case
