+++ vpr/src/base/vpr_types.h
@@ -1449,6 +1449,8 @@ struct t_router_opts {
     bool flat_routing;
     bool has_choking_spot;
 
+    bool with_timing_analysis;
+
     // Options related to rr_node reordering, for testing and possible cache optimization
     e_rr_node_reorder_algorithm reorder_rr_graph_nodes_algorithm = DONT_REORDER;
     int reorder_rr_graph_nodes_threshold = 0;
