+++ vpr/src/draw/draw.cpp
@@ -86,7 +86,7 @@
 #    endif
 
 #    include "rr_graph.h"
-#    include "route_util.h"
+#    include "route_utilization.h"
 #    include "place_macro.h"
 #    include "buttons.h"
 #    include "draw_rr.h"
