+++ vpr/src/draw/search_bar.cpp
@@ -59,7 +59,7 @@
 #    endif
 
 #    include "rr_graph.h"
-#    include "route_util.h"
+#    include "route_utilization.h"
 #    include "place_macro.h"
 
 extern std::string rr_highlight_message;
