+++ vpr/src/route/channel_stats.cpp
@@ -1,5 +1,5 @@
 #include "channel_stats.h"
-#include "route_util.h"
+#include "route_utilization.h"
 #include "histogram.h"
 #include "globals.h"
