+++ vpr/src/route/connection_router_interface.h
@@ -37,7 +37,7 @@ class ConnectionRouterInterface {
   public:
     virtual ~ConnectionRouterInterface() {}
 
-    // Clear's the modified list.  Should be called after reset_path_costs
+    // Clears the modified list.  Should be called after reset_path_costs
     // have been called.
     virtual void clear_modified_rr_node_info() = 0;
 
@@ -50,7 +50,7 @@ class ConnectionRouterInterface {
      *
      * Returns a tuple of:
      * bool: path exists? (hard failure, rr graph disconnected)
-     * bool: should retry with full bounding box? (only used in parallel routing)
+     * bool: should retry with full bounding box?
      * t_heap: heap element of cheapest path */
     virtual std::tuple<bool, bool, t_heap> timing_driven_route_connection_from_route_tree(
         const RouteTreeNode& rt_root,
@@ -58,8 +58,7 @@ class ConnectionRouterInterface {
         const t_conn_cost_params cost_params,
         t_bb bounding_box,
         RouterStats& router_stats,
-        const ConnectionParameters& conn_params,
-        bool can_grow_bb)
+        const ConnectionParameters& conn_params)
         = 0;
 
     /** Finds a path from the route tree rooted at rt_root to sink_node for a
@@ -70,7 +69,7 @@ class ConnectionRouterInterface {
      *
      * Returns a tuple of:
      * bool: path exists? (hard failure, rr graph disconnected)
-     * bool: should retry with full bounding box? (only used in parallel routing)
+     * bool: should retry with full bounding box?
      * t_heap: heap element of cheapest path */
     virtual std::tuple<bool, bool, t_heap> timing_driven_route_connection_from_route_tree_high_fanout(
         const RouteTreeNode& rt_root,
@@ -79,8 +78,7 @@ class ConnectionRouterInterface {
         t_bb bounding_box,
         const SpatialRouteTreeLookup& spatial_rt_lookup,
         RouterStats& router_stats,
-        const ConnectionParameters& conn_params,
-        bool can_grow_bb)
+        const ConnectionParameters& conn_params)
         = 0;
 
     // Finds a path from the route tree rooted at rt_root to all sinks
