+++ vpr/src/route/route_budgets.cpp
@@ -44,7 +44,6 @@
 
 #include "vtr_assert.h"
 #include "vtr_log.h"
-#include "route_timing.h"
 #include "tatum/report/TimingPathFwd.hpp"
 #include "tatum/base/TimingType.hpp"
 #include "concrete_timing_info.h"
