+++ vpr/src/route/route_common.h
@@ -1,14 +1,19 @@
-/************ Defines and types shared by all route files ********************/
 #pragma once
+
+/** @file Misc. router utils: some used by the connection router, some by other
+ * router files and some used globally. */
+
 #include <vector>
 #include "clustered_netlist.h"
-#include "vtr_vector.h"
-#include "heap_type.h"
 #include "rr_node_fwd.h"
 #include "router_stats.h"
 #include "globals.h"
 
-/******* Subroutines in route_common used only by other router modules ******/
+/** This routine checks to see if this is a resource-feasible routing.
+ * That is, are all rr_node capacity limitations respected?  It assumes
+ * that the occupancy arrays are up to date when it is called. */
+bool feasible_routing();
+
 vtr::vector<ParentNetId, t_bb> load_route_bb(const Netlist<>& net_list,
                                              int bb_factor);
 
@@ -22,8 +27,6 @@ void pathfinder_update_acc_cost_and_overuse_info(float acc_fac, OveruseInfo& ove
 
 /** Update pathfinder cost of all nodes under root (including root) */
 void pathfinder_update_cost_from_route_tree(const RouteTreeNode& root, int add_or_sub);
-
-float update_pres_fac(float new_pres_fac);
 
 void reset_path_costs(const std::vector<RRNodeId>& visited_rr_nodes);
