+++ vpr/src/route/route_utilization.cpp
@@ -1,4 +1,4 @@
-#include "route_util.h"
+#include "route_utilization.h"
 #include "globals.h"
 #include "draw_types.h"
 #include "draw_global.h"
