+++ vpr/src/route/router_delay_profiling.h
@@ -2,7 +2,6 @@
 #define ROUTER_DELAY_PROFILING_H_
 
 #include "vpr_types.h"
-#include "route_timing.h"
 #include "binary_heap.h"
 #include "connection_router.h"
