+++ vpr/src/route/router_stats.h
@@ -51,6 +51,30 @@ struct RouterStats {
     size_t add_all_rt_from_high_fanout = 0;
     size_t add_high_fanout_rt = 0;
     size_t add_all_rt = 0;
+
+    /** Add rhs's stats to mine */
+    void combine(RouterStats& rhs) {
+        connections_routed += rhs.connections_routed;
+        nets_routed += rhs.nets_routed;
+        heap_pushes += rhs.heap_pushes;
+        inter_cluster_node_pushes += rhs.inter_cluster_node_pushes;
+        intra_cluster_node_pushes += rhs.intra_cluster_node_pushes;
+        heap_pops += rhs.heap_pops;
+        inter_cluster_node_pops += rhs.inter_cluster_node_pops;
+        intra_cluster_node_pops += rhs.intra_cluster_node_pops;
+        for (int node_type_idx = 0; node_type_idx < t_rr_type::NUM_RR_TYPES; node_type_idx++) {
+            inter_cluster_node_type_cnt_pushes[node_type_idx] += rhs.inter_cluster_node_type_cnt_pushes[node_type_idx];
+            inter_cluster_node_type_cnt_pops[node_type_idx] += rhs.inter_cluster_node_type_cnt_pops[node_type_idx];
+            intra_cluster_node_type_cnt_pushes[node_type_idx] += rhs.intra_cluster_node_type_cnt_pushes[node_type_idx];
+            intra_cluster_node_type_cnt_pops[node_type_idx] += rhs.intra_cluster_node_type_cnt_pops[node_type_idx];
+            rt_node_pushes[node_type_idx] += rhs.rt_node_pushes[node_type_idx];
+            rt_node_high_fanout_pushes[node_type_idx] += rhs.rt_node_high_fanout_pushes[node_type_idx];
+            rt_node_entire_tree_pushes[node_type_idx] += rhs.rt_node_entire_tree_pushes[node_type_idx];
+        }
+        add_all_rt += rhs.add_all_rt;
+        add_all_rt_from_high_fanout += rhs.add_all_rt_from_high_fanout;
+        add_high_fanout_rt += rhs.add_high_fanout_rt;
+    }
 };
 
 class WirelengthInfo {
