+++ vtr_flow/scripts/python_libs/vtr/util.py
@@ -442,6 +442,7 @@ REUSABLE_FILES = {
     "route": ["route", "--route_file"],
     "rr_graph": ["rr_graph.xml", "--read_rr_graph"],
     "lookahead": ["lookahead.bin", "--read_router_lookahead"],
+    "blif": ["pre-vpr.blif", "REPLACE_BLIF"],
 }
 
 
@@ -456,11 +457,12 @@ def argparse_use_previous(inp: str) -> List[Tuple[str, List]]:
     for w in tokens:
         r = re.fullmatch(r"(\w+):(\w+)", w)
         if not r:
-            raise argparse.ArgumentTypeError("Invalid input to -use_previous: %s" % w)
+            raise argparse.ArgumentError(None, "Invalid input to -use_previous: %s" % w)
         if not REUSABLE_FILES.get(r.group(2)):
-            raise argparse.ArgumentTypeError(
+            raise argparse.ArgumentError(
+                None,
                 "Unknown file type to use_previous: %s, available types: %s"
-                % (r.group(2), ",".join(REUSABLE_FILES.keys()))
+                % (r.group(2), ",".join(REUSABLE_FILES.keys())),
             )
         out.append((r.group(1), REUSABLE_FILES[r.group(2)]))
